+++ changelog.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- You can now force load checkpoints in the Checkpoint Manager if there's an error \[[Issue #13](https://github.com/Kevadroz/PracticeCheckpointPermanence/issues/13)\]
 - The attempts are no longer saved in the checkpoints, now the attempt counter is reset to one (configurable) \[[Issue #14](https://github.com/Kevadroz/PracticeCheckpointPermanence/issues/14)\]
 
 ## 1.3.1 ([Release](https://github.com/Kevadroz/PracticeCheckpointPermanence/releases/tag/v1.3.1)) ([Source](https://github.com/Kevadroz/PracticeCheckpointPermanence/tree/v1.3.1))